fix(oc): use the full F35SQB004G — 512 MB flash, not 128 MB (#492)
The rocket computer reported (and used) only 128 MB of its 4 Gbit / 512 MB
SPI NAND. The flight-log geometry constants described a 1 Gbit chip: 2048 B
pages (chip is 4096) and 1024 blocks (chip has 2048). Two compounding 2x
errors meant the raw-SPI path wrote only the first half of every physical
page and never touched the upper 1024 blocks — 3/4 of the chip dark. (The
IDF spi_nand_flash driver had the right geometry but is off the flight
path; TR_LogToFlash's bit-banged path is authoritative.)

Corrected both duplicate geometry blocks to page=4096, blocks=2048, extended
the flight region + metadata into the upper half, and halved prealloc/extend
so the ~10 MB / ~200 ms byte-intent holds at the now-256 KB block. Widened
RocketStorageStatsData.block_size_kb to u16 (256 overflowed u8 -> 0), giving
a 15-byte frame, and updated the app decoder to match.

Stack safety (page buffers double 2 KB -> 4 KB): writeFrame's page buffer plus
the nested retireBlockAndSalvage buffer would have put 8 KB on the 8 KB flush
task stack during mid-flight bad-block relocation. Moved both to .bss member
scratch buffers (flightlog is a global; all paths are serialized by mutex_).
The boot MRAM->LFS recovery buffer reuses the idle staging member so it no
longer stacks with the LFS read callback.

DESTRUCTIVE MIGRATION: changing page size + block count + metadata location
reformats LittleFS, reseeds the NVS block bitmaps, and relocates the flight
index — existing stored flights are lost. Download before flashing.

Bench-verify: 512 MB reported in-app; logging + read-back at 4080 B/page;
oc_loop/flush stack high-water margin; F35SQB004G OOB bad-block marker column
(now 4096) and full-page on-die ECC.

+++ tinkerrocket-idf/projects/out_computer/main/main.cpp
@@ -99,7 +99,7 @@ static TR_LogToFlash logger;
 // writes via writeFrame(), with dual-copy index metadata in blocks 1020-1023
 // and a persistent 3-state bitmap in NVS. TR_LogToFlash keeps the ring/flush
 // machinery and shelled-down 4 MB LFS partition for config; a write_sink
-// fn-pointer routes each drained 2032 B page into flightlog.writeFrame().
+// fn-pointer routes each drained 4080 B page into flightlog.writeFrame().
 static tr_flightlog::TR_NandBackend_esp flightlog_backend;
 static tr_flightlog::TR_FlightLog flightlog;
 static tr_flightlog::NvsBitmapStore flightlog_bitmap_store;
@@ -146,7 +146,7 @@ static SensorHealthState ocStorageHealth()
 
 // Stage 3b (issue #50): BLE file-ops re-backed on TR_FlightLog.
 // Fills up to `max_bytes` of the downloader's scratch buffer by issuing
-// successive readFlightPage calls (each one returns at most 2032 payload
+// successive readFlightPage calls (each one returns at most 4080 payload
 // bytes — the portion of a single NAND page past its PageHeader). Matches
 // the logger.readFileChunk contract: sets eof=true when the last byte of
 // the flight has been read; returns false only on underlying NAND I/O error.
@@ -3908,7 +3908,7 @@ static void printStats()
                 rss.bad_blocks    = (uint16_t)flightlog.bitmap().countInState(tr_flightlog::BLOCK_BAD);
                 rss.system_blocks = (uint16_t)(fcfg.flight_region_start + 4u);  // LFS region + 4 metadata
                 rss.flight_count  = (uint16_t)flightlog.index().size();
-                rss.block_size_kb = (uint8_t)(tr_flightlog::NAND_BLOCK_SIZE / 1024u);
+                rss.block_size_kb = (uint16_t)(tr_flightlog::NAND_BLOCK_SIZE / 1024u);
                 rss.flags         = 0x01;  // initialized
             }
             ble_app.sendStorageStats(0xCC, reinterpret_cast<const uint8_t*>(&rss), sizeof(rss));
@@ -4386,7 +4386,7 @@ void initPeripherals()
 
     // --- LFS shrunk to 4 MB + hot-path write sink (issue #50) ---------------
     // LFS holds 32 blocks for config/placeholder use; TR_FlightLog owns the
-    // remaining 988 blocks plus the metadata blocks 1020-1023. Each 2032 B
+    // remaining 2012 blocks plus the metadata blocks 2044-2047. Each 4080 B
     // chunk the flush task drains from the ring is routed through
     // flightlogWriteSink → flightlog.writeFrame(), which wraps it in a
     // PageHeader (CRC32 + seq + flight_id) and programs one NAND page
